Contactリスト削除機能の修正（未対応は削除できないようにする）

+++ server/app/Http/Controllers/AdminController.php
@@ -40,9 +40,14 @@ class AdminController extends Controller
 
     public function destroy(Contact $id)
     {
-        $id->delete();
+        if ($id->status == '2') {
+            $id->delete();
 
-        return redirect()->route('contact.list')
-            ->with('status', '削除しました。');
+            return redirect()->route('contact.list')
+                ->with('status', '削除しました。');
+        } else {
+            return redirect()->route('contact.list')
+                ->with('status', '未対応は削除できません。');
+        }
     }
 }
